+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Revision history for antigen
 
+## 0.1.3.0
+
+* Rename `antiNum` and `antiBool` to `fickleNum` and `fickleBool`
+* Rename `antiTry` and `antiTryGen` to `fickleTry` and `fickleTryGen`
+
 ## 0.1.2.0
 
 * Fixed `sized` and `resize` implementations for `AntiGen`